Event is inserting correctly

+++ Models/event.go
@@ -27,7 +27,7 @@ type Event struct {
 	EventEndType                   string                 `bson:"EventEndType" json:"eventEndType"`
 	ClientType                     string                 `bson:"ClientType" json:"clientType"`
 	PeriodTimeForMiddleJoinTillEnd int64                  `bson:"PeriodTimeForMiddleJoinTillEnd" json:"periodTimeForMiddleJoinTillEnd"`
-	ConfigVersion                  int                    `bson:"ConfigVersion" json:"configVersion"`
+	ConfigVersion                  float64                    `bson:"ConfigVersion" json:"configVersion"`
 	States                         map[string]interface{} `bson:"States" json:"states"`
 	VersionMetaData                map[string]interface{} `bson:"VersionMetaData" json:"versionMetaData"`
 	Repetition                     []Repetition           `bson:"Repetition" json:"repetition"`
